First commit
Oscilloscope display using SSD1306 and Arduino. Also uses the Adafruit OLED and SSD1306 libraries. There is an analogue circuit that needs to be used to offset negative input voltages.
This is the first draft. The basics work but it needs to be tweaked:
Waveforms are inconsistent.
I intend to add some kind of menu structure.
Long term goal is to port this to an STM32f4 as it is more powerful.